packages/qiskit/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning", "setuptools"]
build-backend = "poetry_dynamic_versioning.backend"



[tool.poetry]
name = "jijmodeling-transpiler-quantum-qiskit"
version = "0.1.0rc2"
description = "JijModeling Transpiler for Qiskit"
authors = ["Jij Inc."]
readme = "README.md"
packages = [{include = "jijmodeling_transpiler_quantum"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
qiskit = ">=0.43.1"
jijmodeling-transpiler-quantum = "*"


[tool.poetry.group.dev.dependencies]
matplotlib = "^3.7.2"
jijmodeling-transpiler-quantum = {path = "../../", develop = true}
